requirements.txt
numpy
python-mnist